Removed tests module
