+++ encoder-otel-brave/src/test/java/zipkin2/reporter/otel/brave/JaegerAllInOne.java
@@ -11,11 +11,10 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package zipkin2.reporter.otel;
+package zipkin2.reporter.otel.brave;
 
 import java.util.Collections;
 import java.util.Set;
-
 import org.testcontainers.containers.GenericContainer;
 import org.testcontainers.containers.wait.strategy.HttpWaitStrategy;
